test: Trajectory time array is now float64 (was float32)
molrs 0.1.5's Trajectory binding requires a float64 'time' array; passing
float32 raised a misleading 'ndarray is not an instance of ndarray'. Match
the updated dtype. Full molrec suite now 56 passed.

+++ src/molrec/tests/test_molrec_trajectory.py
@@ -11,7 +11,7 @@ class TestTrajectoryFromPDB:
         trajectory = molrs.Trajectory(
             [frame0, frame1],
             step=np.array([0, 1], dtype=np.int64),
-            time=np.array([0.0, 1.0], dtype=np.float32),
+            time=np.array([0.0, 1.0], dtype=np.float64),
         )
         record = molrs.MolRec()
         record.set_frame(frame0)
